Fix preview of theming
Pull request #5429 made cached SCSS files depend on a hash of the base
URL, so the "/css/core/server.css" file does no longer exist; as the
file can not be loaded the "Loading preview" message is never removed
and the "Saved" message is never shown.

As it now depends on the hash of the base URL the file to be reloaded
can no longer be hardcoded, so the DataResponse from the controller now
provides the full URL to the "server.css" file that has to be reloaded
(if any).

Fixes #5975

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>